backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
Pillow==11.3.0
psycopg2-binary
python-dotenv
websockets
